TYP: libgroupby int64->intp

+++ pandas/tests/test_algos.py
@@ -1712,7 +1712,7 @@ def test_quantile():
 
 def test_unique_label_indices():
 
-    a = np.random.randint(1, 1 << 10, 1 << 15).astype("i8")
+    a = np.random.randint(1, 1 << 10, 1 << 15).astype("int64")
 
     left = ht.unique_label_indices(a)
     right = np.unique(a, return_index=True)[1]
